move config to workflow file
